requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
werkzeug==3.0.1
pandas==2.2.0
nltk==3.8.1
gensim==4.3.2
Sastrawi==1.0.1
